created a hook for POST /reports/generate to be used in the reports page
